Added helpful utils
